feat(nav header): create nav header component
