+++ src/shared/nav-header/index.ts
@@ -1,0 +1,1 @@
+export { NavHeader } from './NavHeader';
