+++ src/shared/nav-header/navHeader.module.css
@@ -1,0 +1,13 @@
+.routeDepthContainer {
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	padding-left: 1.75rem;
+	padding-right: 1.75rem;
+	padding-top: 10px;
+	padding-bottom: 10px;
+	min-height: 40px;
+	top: 0;
+	background-color: #fff;
+	box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.05);
+}
